requirements.txt
alembic==1.14.0
blinker==1.9.0
cffi==1.17.1
click==8.1.7
colorama==0.4.6
commonmark==0.9.1
cryptography==43.0.3
Deprecated==1.2.14
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Limiter==2.7.0
Flask-Login==0.6.3
Flask-Migrate==4.0.7
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
greenlet==3.1.1
importlib_resources==6.4.5
itsdangerous==2.2.0
Jinja2==3.1.4
limits==3.13.0
Mako==1.3.6
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==24.2
pyClamd==0.4.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.6.0
PyMySQL==1.0.2
python-dotenv==1.0.1
python-magic==0.4.27
rich==12.6.0
SQLAlchemy==2.0.36
typing_extensions==4.12.2
Werkzeug==3.1.3
wrapt==1.16.0
WTForms==3.2.1
